# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Crm",
        "description": "crm vue jwt",
        "contact": {
            "email": "info@digex"
        },
        "license": {
            "name": "crm",
            "url": "http://crm.loc/"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "",
            "description": "crm  host server"
        }
    ],
    "paths": {
...
        "/api/incomes/categories/destroy/{item}": {
            "delete": {
                "tags": [
                    "incomes"
                ],
                "description": "get company by id",
                "operationId": "App\\Http\\Controllers\\InvoiceCategoryController::destroy",
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "description": "id of category",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        },
...
        "/api/persons/index": {
            "get": {
                "tags": [
                    "persons"
                ],
                "description": "get companies",
                "operationId": "App\\Http\\Controllers\\PersonController::index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "number of page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        },
...
        "/api/settings/taxes/index": {
            "get": {
                "tags": [
                    "settings"
                ],
                "description": "get companies",
                "operationId": "App\\Http\\Controllers\\TaxController::index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "number of page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        },
        "/api/settings/taxes/get": {
            "get": {
                "tags": [
                    "settings"
                ],
                "description": "get companies",
                "operationId": "App\\Http\\Controllers\\TaxController::get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id of category",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "auth": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "auth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}